requirements.txt
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cssselect==1.1.0
git+git://github.com/Rapptz/discord.py.git
fake-useragent==0.1.11
flameprof==0.4
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
importlib-metadata==4.8.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.2.0
parse==1.19.0
psutil==5.8.0
psycopg2==2.9.3
pyee==8.2.2
pyppeteer==0.2.6
pypresence==4.2.1
pyquery==1.4.3
python-dotenv==0.19.2
requests==2.26.0
schedule==1.1.0
six==1.16.0
smmap==3.0.5
soupsieve==2.2.1
tqdm==4.62.2
typing_extensions==4.0.0
w3lib==1.22.0
websockets==9.1
Werkzeug==1.0.1
xxhash==1.4.4
yarl==1.7.2
zipp==3.5.0

